# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "DatasetStatus": {
        "additionalProperties": false,
        "properties": {
          "dataset": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dataset"
          },
          "dataset_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dataset Version"
          },
          "generated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generated At"
          },
          "license_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "License Note"
          },
          "play_count": {
            "title": "Play Count",
            "type": "integer"
          },
          "schema_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schema Version"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "subset": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subset"
          },
          "synthetic": {
            "title": "Synthetic",
            "type": "boolean"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "required": [
          "dataset",
          "title",
          "source",
          "subset",
          "dataset_version",
          "schema_version",
          "play_count",
          "synthetic",
          "license_note",
          "generated_at"
        ],
        "title": "DatasetStatus",
        "type": "object"
      },
...
      "ErrorEnvelope": {
        "additionalProperties": false,
        "description": "Every non-2xx response has this shape.",
        "example": {
          "error": {
            "code": "play_not_found",
            "message": "Play 2023090700-999 is not in dataset nfl_bdb_2026_analytics@dev-46b081804b38.",
            "request_id": "5f2c9a1be0d4",
            "status": 404
          }
        },
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
      "Facets": {
        "additionalProperties": false,
        "properties": {
          "coverages": {
            "items": {
              "type": "string"
            },
            "title": "Coverages",
            "type": "array"
          },
          "formations": {
            "items": {
              "type": "string"
            },
            "title": "Formations",
            "type": "array"
          },
          "play_types": {
            "items": {
              "enum": [
                "pass",
                "run",
                "other"
              ],
              "type": "string"
            },
            "title": "Play Types",
            "type": "array"
          },
          "quarters": {
            "items": {
              "type": "integer"
            },
            "title": "Quarters",
            "type": "array"
          },
          "seasons": {
            "items": {
              "type": "integer"
            },
            "title": "Seasons",
            "type": "array"
          },
          "teams": {
            "items": {
              "type": "string"
            },
            "title": "Teams",
            "type": "array"
          },
          "weeks": {
            "items": {
              "type": "integer"
            },
            "title": "Weeks",
            "type": "array"
          }
        },
        "required": [
          "seasons",
          "weeks",
          "teams",
          "formations",
          "coverages",
          "play_types",
          "quarters"
        ],
        "title": "Facets",
        "type": "object"
      },
...
      "FramesPayload": {
        "additionalProperties": false,
        "description": "Observed tracking only. Never contains future (output-file) positions.",
        "properties": {
          "coordinate_system": {
            "title": "Coordinate System",
            "type": "string"
          },
          "dataset_version": {
            "title": "Dataset Version",
            "type": "string"
          },
          "frame_rate_hz": {
            "title": "Frame Rate Hz",
            "type": "number"
          },
          "frames": {
            "items": {
              "$ref": "#/components/schemas/Frame"
            },
            "title": "Frames",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "observed_frame_count": {
            "title": "Observed Frame Count",
            "type": "integer"
          },
          "schema_version": {
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "id",
          "dataset_version",
          "schema_version",
          "coordinate_system",
          "frame_rate_hz",
          "observed_frame_count",
          "frames"
        ],
        "title": "FramesPayload",
        "type": "object"
      },
...
      "PlayDetail": {
        "additionalProperties": false,
        "properties": {
          "annotations": {
            "$ref": "#/components/schemas/PlayAnnotations"
          },
          "away_team": {
            "title": "Away Team",
            "type": "string"
          },
          "ball_landing": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BallLanding"
              },
              {
                "type": "null"
              }
            ]
          },
          "context": {
            "$ref": "#/components/schemas/PlayContext"
          },
          "coordinates": {
            "$ref": "#/components/schemas/CoordinateInfo"
          },
          "defense": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Defense"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Supplied narrative; describes the result (post-play).",
            "title": "Description"
          },
          "down": {
            "anyOf": [
              {
                "maximum": 4.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Down"
          },
          "events": {
            "description": "Supplied event annotations; this dataset has none.",
            "items": {
              "$ref": "#/components/schemas/PlayEvent"
            },
            "title": "Events",
            "type": "array"
          },
          "first_down_x": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Down X"
          },
          "frame_rate_hz": {
            "title": "Frame Rate Hz",
            "type": "number"
          },
          "game_clock": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Game Clock"
          },
          "game_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "ISO date.",
            "title": "Game Date"
          },
          "game_id": {
            "title": "Game Id",
            "type": "integer"
          },
          "home_team": {
            "title": "Home Team",
            "type": "string"
          },
          "id": {
            "description": "PlayLens play ID '<game_id>-<play_id>'.",
            "examples": [
              "2023091008-3826"
            ],
            "title": "Id",
            "type": "string"
          },
          "line_of_scrimmage_x": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Line Of Scrimmage X"
          },
          "offense": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Offense"
          },
          "outcome": {
            "$ref": "#/components/schemas/PlayOutcome"
          },
          "play_direction": {
            "anyOf": [
              {
                "enum": [
                  "left",
                  "right"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Recorded attack direction. Coordinates are canonical.",
            "title": "Play Direction"
          },
          "play_id": {
            "description": "NFL play ID; unique only within its game.",
            "title": "Play Id",
            "type": "integer"
          },
          "play_type": {
            "anyOf": [
              {
                "enum": [
                  "pass",
                  "run",
                  "other"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Play Type"
          },
          "players": {
            "items": {
              "$ref": "#/components/schemas/PlayerRef"
            },
            "title": "Players",
            "type": "array"
          },
          "provenance": {
            "$ref": "#/components/schemas/Provenance"
          },
          "quarter": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quarter"
          },
          "season": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Season"
          },
          "tracking": {
            "$ref": "#/components/schemas/TrackingSummary"
          },
          "week": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Week"
          },
          "yardline_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "e.g. 'DET 41'; '50' at midfield.",
            "title": "Yardline Label"
          },
          "yards_to_go": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Yards To Go"
          }
        },
        "required": [
          "id",
          "game_id",
          "play_id",
          "season",
          "week",
          "game_date",
          "home_team",
          "away_team",
          "offense",
          "defense",
          "quarter",
          "game_clock",
          "down",
          "yards_to_go",
          "yardline_label",
          "play_type",
          "description",
          "context",
          "annotations",
          "outcome",
          "tracking",
          "players",
          "events",
          "play_direction",
          "line_of_scrimmage_x",
          "first_down_x",
          "frame_rate_hz",
          "ball_landing",
          "coordinates",
          "provenance"
        ],
        "title": "PlayDetail",
        "type": "object"
      },
...
      "PlayPage": {
        "additionalProperties": false,
        "description": "One page of plays, newest game first.",
        "properties": {
          "dataset": {
            "$ref": "#/components/schemas/DatasetStatus"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/PlaySummary"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "similarity": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SimilarityScores"
              },
              {
                "type": "null"
              }
            ],
            "description": "Always null: no retrieval model is served."
          },
          "sort": {
            "enum": [
              "recent",
              "similarity"
            ],
            "title": "Sort",
            "type": "string"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "page_size",
          "sort",
          "similarity",
          "dataset"
        ],
        "title": "PlayPage",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "\nReal NFL tracking data (NFL Big Data Bowl 2026 Analytics) in the PlayLens canonical\ncoordinate frame. Observed frames, held-out future trajectories, and post-play\nmetadata are separate resources. No model is served in this version.\n",
    "title": "PlayLens API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/dataset": {
      "get": {
        "operationId": "dataset_api_v1_dataset_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Processed dataset not loaded."
          }
        },
        "summary": "Dataset served by this API",
        "tags": [
          "plays"
        ]
      }
    },
...
    "/api/v1/plays": {
      "get": {
        "operationId": "list_plays_api_v1_plays_get",
        "parameters": [
          {
            "description": "Case-insensitive search over ID, teams, description, formation, coverage, and route.",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Case-insensitive search over ID, teams, description, formation, coverage, and route.",
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "season",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Season"
            }
          },
          {
            "in": "query",
            "name": "week",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 22,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Week"
            }
          },
          {
            "in": "query",
            "name": "offense",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Offense"
            }
          },
          {
            "in": "query",
            "name": "defense",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Defense"
            }
          },
          {
            "description": "offense_formation, e.g. SHOTGUN.",
            "in": "query",
            "name": "formation",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "offense_formation, e.g. SHOTGUN.",
              "title": "Formation"
            }
          },
          {
            "description": "team_coverage_type, e.g. COVER_3_ZONE.",
            "in": "query",
            "name": "coverage",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "team_coverage_type, e.g. COVER_3_ZONE.",
              "title": "Coverage"
            }
          },
          {
            "in": "query",
            "name": "down",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 4,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Down"
            }
          },
          {
            "description": "1-3, 4-7, 8+ yards to go.",
            "in": "query",
            "name": "distance",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "short",
                    "medium",
                    "long"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "1-3, 4-7, 8+ yards to go.",
              "title": "Distance"
            }
          },
          {
            "in": "query",
            "name": "quarter",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 5,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Quarter"
            }
          },
          {
            "in": "query",
            "name": "play_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "pass",
                    "run",
                    "other"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Play Type"
            }
          },
          {
            "description": "By yards_gained.",
            "in": "query",
            "name": "outcome",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "gain",
                    "no_gain",
                    "loss",
                    "unknown"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "By yards_gained.",
              "title": "Outcome"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "recent",
              "enum": [
                "recent",
                "similarity"
              ],
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "similar_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Similar To"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Malformed play ID or query."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Processed dataset not loaded."
          }
        },
        "summary": "Browse plays",
        "tags": [
          "plays"
        ]
      }
    },
    "/api/v1/plays/facets": {
      "get": {
        "operationId": "facets_api_v1_plays_facets_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Facets"
                }
              }
            },
            "description": "Successful Response"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Processed dataset not loaded."
          }
        },
        "summary": "Filter values present in the dataset",
        "tags": [
          "plays"
        ]
      }
    },
    "/api/v1/plays/{play_id}": {
      "get": {
        "operationId": "get_play_api_v1_plays__play_id__get",
        "parameters": [
          {
            "description": "PlayLens play ID '<game_id>-<play_id>'.",
            "in": "path",
            "name": "play_id",
            "required": true,
            "schema": {
              "description": "PlayLens play ID '<game_id>-<play_id>'.",
              "examples": [
                "2023091008-3826"
              ],
              "title": "Play Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unknown play."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Malformed play ID or query."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Processed dataset not loaded."
          }
        },
        "summary": "Play metadata and roster",
        "tags": [
          "plays"
        ]
      }
    },
    "/api/v1/plays/{play_id}/frames": {
      "get": {
        "operationId": "get_frames_api_v1_plays__play_id__frames_get",
        "parameters": [
          {
            "description": "PlayLens play ID '<game_id>-<play_id>'.",
            "in": "path",
            "name": "play_id",
            "required": true,
            "schema": {
              "description": "PlayLens play ID '<game_id>-<play_id>'.",
              "examples": [
                "2023091008-3826"
              ],
              "title": "Play Id",
              "type": "string"
            }
          },
          {
            "description": "First frame_id to include.",
            "in": "query",
            "name": "start_frame",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "First frame_id to include.",
              "title": "Start Frame"
            }
          },
          {
            "description": "Last frame_id to include.",
            "in": "query",
            "name": "end_frame",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Last frame_id to include.",
              "title": "End Frame"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FramesPayload"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unknown play."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Malformed play ID or query."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Processed dataset not loaded."
          }
        },
        "summary": "Observed tracking frames",
        "tags": [
          "plays"
        ]
      }
    },
    "/api/v1/plays/{play_id}/future": {
      "get": {
        "operationId": "get_future_api_v1_plays__play_id__future_get",
        "parameters": [
          {
            "description": "PlayLens play ID '<game_id>-<play_id>'.",
            "in": "path",
            "name": "play_id",
            "required": true,
            "schema": {
              "description": "PlayLens play ID '<game_id>-<play_id>'.",
              "examples": [
                "2023091008-3826"
              ],
              "title": "Play Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FuturePayload"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unknown play."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Malformed play ID or query."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Processed dataset not loaded."
          }
        },
        "summary": "Held-out actual future trajectories (ground truth, not predictions)",
        "tags": [
          "plays"
        ]
      }
    },
...
  }
}